pyproject.toml
[project]
name = "chat-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "chromadb>=1.4.1",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.6",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
